# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeBattle Arena API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  page: number;
  totalPages: number;
}