# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeBattle Arena API
 * OpenAPI spec version: 0.1.0
 */
import type { Example } from './example';
import type { ProblemDifficulty } from './problemDifficulty';
import type { ProblemStarterCode } from './problemStarterCode';

export interface Problem {
  id: number;
  title: string;
  description: string;
  difficulty: ProblemDifficulty;
  /** @nullable */
  constraints?: string | null;
  examples?: Example[];
  tags: string[];
  starterCode: ProblemStarterCode;
  solvedBy?: number;
}