# lib/api-zod/src/generated/types/problemListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeBattle Arena API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemSummary } from './problemSummary';

export interface ProblemListResponse {
  problems: ProblemSummary[];
  total: number;
  page: number;
  totalPages: number;
}