# lib/api-zod/src/generated/types/problemSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeBattle Arena API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemSummaryDifficulty } from './problemSummaryDifficulty';

export interface ProblemSummary {
  id: number;
  title: string;
  difficulty: ProblemSummaryDifficulty;
  tags: string[];
  solvedBy: number;
}