# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeBattle Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  rating: number;
  wins: number;
  losses: number;
  rank: string;
  solvedCount?: number;
  createdAt: string;
}